requirements.txt
openai==1.3.3
azure-cognitiveservices-speech==1.36.0
spotipy==2.23.0
google-cloud-texttospeech==2.16.3
pyaudio==0.2.14
numpy==1.26.3
aiohttp==3.9.3
sqlalchemy==2.0.29
backoff==2.2.1
